+++ lib/pbio/drv/counter/counter_nxt.c
@@ -35,7 +35,7 @@ static const pbdrv_counter_funcs_t pbdrv_counter_nxt_funcs = {
 };
 
 void pbdrv_counter_nxt_init(pbdrv_counter_dev_t *devs) {
-    for (int i = 0; i < PBIO_ARRAY_SIZE(private_data); i++) {
+    for (size_t i = 0; i < PBIO_ARRAY_SIZE(private_data); i++) {
         private_data_t *priv = &private_data[i];
 
         priv->port = i;
