+++ lib/pbio/sys/light_matrix.c
@@ -80,7 +80,7 @@ static uint32_t pbsys_hub_light_matrix_user_program_animation_next(pbio_light_an
 
     pbdrv_led_array_dev_t *array;
     if (pbdrv_led_array_get_dev(0, &array) == PBIO_SUCCESS) {
-        for (int i = 0; i < PBIO_ARRAY_SIZE(indexes); i++) {
+        for (size_t i = 0; i < PBIO_ARRAY_SIZE(indexes); i++) {
             // The pixels are spread equally across the pattern.
             uint8_t offset = cycle + i * (UINT8_MAX / PBIO_ARRAY_SIZE(indexes));
             uint8_t brightness = offset > 200 ? 0 : (offset < 100 ? offset : 200 - offset);
